agents/cartao_credito/requirements.txt
fastapi
uvicorn
langchain_openai
langchain-core
pydantic
python-dotenv
httpx
a2a-sdk[http-server]
